added passport authentication

+++ routes/posts.js
@@ -3,25 +3,27 @@ var Post = require('./../models/post')
 var router = express.Router();
 // var mongoose = require('mongoose');
 
-router.get('/new', (req, res) => {
-    res.render('new', { post: new Post() });
+router.get('/new', checkAuthenticated, (req, res) => {
+    res.render('new', { post: new Post()});
 })
 
-router.get('/edit/:id', async (req, res) => {
+router.get('/edit/:id', checkAuthenticated, async (req, res) => {
     const post = await Post.findById(req.params.id)
     res.render('edit', {post: post})
 })
 
-router.get('/:id', async (req, res) => {
+router.get('/:id', checkAuthenticated, async (req, res) => {
     let post = await Post.findById({_id: req.params.id})
     if(post == null) res.redirect('/')
-    res.render('show', {post: post})
+    res.render('show', {post: post, user: req.user})
 })
 
-router.post('/', async (req, res) => {
+router.post('/', checkAuthenticated, async (req, res) => {
+
     let post = new Post({
         title: req.body.title,
-        description: req.body.description
+        description: req.body.description,
+        user: req.user.id
     })
     try{
         await post.save()
@@ -33,28 +35,50 @@ router.post('/', async (req, res) => {
     
 });
 
-router.delete('/:id', async (req, res) => {
-    await Post.findByIdAndDelete(req.params.id)
-    res.redirect('/')
+router.delete('/:id', checkAuthenticated, async (req, res) => {
+    let post = await Post.findById(req.params.id)
+    if((post.user).equals(req.user.id)){
+        await Post.findByIdAndDelete(req.params.id)
+        res.redirect('/')
+    }else{
+        res.json({"auth": "Not Authorized"})
+    }
+    
 })
 
-router.put('/:id', async (req, res) => {
+router.put('/:id', checkAuthenticated, async (req, res) => {
     post = req.body
     post.title = req.body.title
     post.description = req.body.description
 
     
     let update = await Post.findById(req.params.id)
+    if((update.user).equals(req.user.id)){
+        await Post.updateOne({_id: req.params.id}, post)
 
-    await Post.updateOne({_id: req.params.id}, post)
-
-    try{
-        update = await Post.save()
-        console.log('success')
-        res.redirect(`/posts/${req.params.id}`)
-    }catch(e){
-        res.redirect(`/posts/${req.params.id}`)
+        try{
+            update = await Post.save()
+            console.log('success')
+            res.redirect(`/posts/${req.params.id}`)
+        }catch(e){
+            res.redirect(`/posts/${req.params.id}`)
+        }
+    }else{
+        res.json({"auth": "Not Authorized"})
     }
+
+
+
+    
 })
+
+function checkAuthenticated(req, res, next){
+  if(req.isAuthenticated()){
+    return next()
+  }
+
+  res.redirect('/login')
+}
+
 
 module.exports = router;
